move to v1/plugin/login-token
