+++ src/v1/index.js
@@ -2,7 +2,7 @@ module.exports = {
     name: 'v1-routes',
     version: '1.0.0',
     register: (server, options) => {
-        server.register(require('./login-tokens'), {
+        server.register(require('./plugin/login-token'), {
             routes: {
                 prefix: '/plugin/login-token'
             }
